% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{numextract}
\alias{numextract}
\title{numextract}
\usage{
numextract(string, return = c("min", "max"))
}
\arguments{
\item{return}{}
}
\value{

}
\description{
numextract
}
\examples{

\dontrun{data(all_cfr)

NE_cfr <- all_cfr \%>\%
dplyr::mutate(min_parts = purrr::map(parts, numextract, "min"),
max_parts = purrr::map(parts, numextract, "max")) \%>\%
dplyr::filter(grepl("Chapter VI", chapters),
min_parts <= 648,
max_parts > 648)}

}
\keyword{internal}
