% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{section_function}
\alias{section_function}
\title{section_function}
\usage{
section_function(data, subpart_number)
}
\arguments{
\item{subpart_number}{}
}
\description{
section_function
}
\keyword{internal}
