% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cfr_part}
\alias{cfr_part}
\title{cfr_part}
\usage{
cfr_part(url, verbose = FALSE)
}
\arguments{
\item{url}{}
}
\description{
cfr_part
}
\examples{

\dontrun{data(cfr_url_list)

cfr_part_list <- cfr_url_list \%>\%
dplyr::filter(!is.na(url)) \%>\%
tidyr::unnest() \%>\%
dplyr::mutate(all_cfr = purrr::map(url, cfr_part))}


}
