% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cfr_text}
\alias{cfr_text}
\title{cfr_text}
\usage{
cfr_text(year, title_number, chapter, part, subpart = NULL,
  return_tidytext = TRUE, verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\description{
cfr_text
}
