% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cfr_urls}
\alias{cfr_urls}
\title{cfr_urls}
\usage{
cfr_urls(year, title, check_url = TRUE, verbose = FALSE)
}
\arguments{
\item{check_url}{}
}
\description{
cfr_urls
}
\examples{

\dontrun{cfr_url_list <- expand.grid(years = 2017:2018,
title = 50,
KEEP.OUT.ATTRS = FALSE,
stringsAsFactors = FALSE) \%>\%
mutate(url = purrr::map2(years, title, cfr_urls, check_url = TRUE))
head(cfr_url_list)}


}
