% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedregs.r
\docType{package}
\name{fedregs}
\alias{fedregs}
\alias{fedregs-package}
\title{\code{fedregs} package}
\description{
Text Analysis of the US Code of Federal Regulations
}
\details{
See the README on
\href{https://github.com/slarge/fedregs/blob/master/README.md}{GitHub}
}
